Revert "Improve node search"
