+++ package.json
@@ -28,7 +28,6 @@
     "@mui/styles": "^5.2.3",
     "@pixi/filter-drop-shadow": "^4.1.5",
     "@pixi/text-html": "^2.0.0",
-    "autosuggest-highlight": "^3.3.4",
     "color": "^4.0.1",
     "core-js": "^3.25.1",
     "cors": "^2.8.5",
@@ -44,7 +43,6 @@
     "lodash": "^4.17.21",
     "lodash-contrib": "^4.1200.1",
     "marked": "^4.1.1",
-    "match-sorter": "^6.3.1",
     "monaco-editor": "^0.34.0",
     "monaco-editor-webpack-plugin": "^7.0.1",
     "notistack": "^2.0.3",
