+++ src/components/GraphOverlayNodeMenu.tsx
@@ -38,7 +38,7 @@ const GraphOverlayNodeMenu: React.FunctionComponent<
             selectionPos.x +
             props.currentGraph.selection.selectionGraphics.width / 2
           }
-          y={Math.max(0, selectionPos.y - 56)}
+          y={Math.max(0, selectionPos.y - 40)}
           selectedNodes={props.selectedNodes}
           randomMainColor={props.randomMainColor}
         />
